requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.9.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
docutils==0.19
exceptiongroup==1.2.0
Faker==20.1.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kitchen==1.2.6
MarkupSafe==2.1.3
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
PyAutoGUI==0.9.54
PyAutoIt==0.6.5
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyobjc-core==10.1
pyobjc-framework-ApplicationServices==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-CoreText==10.1
pyobjc-framework-Quartz==10.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
requests==2.31.0
robotframework==6.0.2
robotframework-appiumlibrary==2.0.0
rubicon-objc==0.4.7
Sanity==0.2.5
selenium==4.15.2
self==2020.12.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
webdriver-manager==4.0.1
Werkzeug==3.0.1
wsproto==1.2.0
